test(farm): align growth stage mock types

+++ backend/internal/farm/catalog_service_test.go
@@ -48,10 +48,10 @@ func TestCatalogServiceListGrowthStagesOrdersBySequence(t *testing.T) {
 
 	s := NewCatalogService(pool)
 	cropID := "00000000-0000-0000-0000-000000000001"
-	pool.ExpectQuery(`SELECT id, crop_id, name, sequence_no, min_days, max_days, COALESCE\(description, ''\) FROM crop_growth_stages WHERE crop_id = \$1 ORDER BY sequence_no`).
+	pool.ExpectQuery(`SELECT id, crop_id, name, sequence_no, min_days::bigint, max_days::bigint, COALESCE\(description, ''\) FROM crop_growth_stages WHERE crop_id = \$1 ORDER BY sequence_no`).
 		WithArgs(cropID).
 		WillReturnRows(pgxmock.NewRows([]string{"id", "crop_id", "name", "sequence_no", "min_days", "max_days", "description"}).
-			AddRow("00000000-0000-0000-0000-000000000003", cropID, "Vegetatif", 1, int32(0), int32(30), "Awal pertumbuhan"))
+			AddRow("00000000-0000-0000-0000-000000000003", cropID, "Vegetatif", 1, int64(0), int64(30), "Awal pertumbuhan"))
 
 	got, err := s.ListGrowthStages(context.Background(), cropID)
 	if err != nil { t.Fatalf("ListGrowthStages() error = %v", err) }
